Edited save for Containers/Filter moved
