+++ src/main/java/de/sbtab/containers/SBTabSpeciesWrapper.java
@@ -28,7 +28,7 @@ public class SBTabSpeciesWrapper implements SBTabElement {
 	}
 
 	@Override
-	public void savaData() {
+	public void saveData() {
 		species.setName(speciesName.getValueSafe());
 		species.setId(speciesId.getValueSafe());
 		species.setCompartment(speciesCompartment.getValueSafe());
@@ -48,6 +48,7 @@ public class SBTabSpeciesWrapper implements SBTabElement {
 
 	public void setSpeciesName(StringProperty speciesName) {
 		this.speciesName = speciesName;
+		saveData();
 	}
 
 	public StringProperty getSpeciesId() {
@@ -56,6 +57,7 @@ public class SBTabSpeciesWrapper implements SBTabElement {
 
 	public void setSpeciesId(StringProperty speciesId) {
 		this.speciesId = speciesId;
+		saveData();
 	}
 
 	public StringProperty getSpeciesCompartment() {
@@ -64,5 +66,6 @@ public class SBTabSpeciesWrapper implements SBTabElement {
 
 	public void setSpeciesCompartment(StringProperty speciesCompartment) {
 		this.speciesCompartment = speciesCompartment;
+		saveData();
 	}
 }
